fix: SleepLogsをContext化 feat: SleepLogsをContext化 #13
削除機能以外でもsetできるように、Context化しました。
