+++ front/react-app/src/Organisms/SleepLogManipulate.tsx
@@ -1,25 +1,28 @@
 import axios from 'axios';
-import { useCallback, useRef, useState } from 'react';
+import { useCallback, useContext, useRef, useState } from 'react';
 import { VscChromeClose, VscEllipsis, VscEdit, VscTrash } from 'react-icons/vsc';
 import { sleepLogsAPI } from '../constants/urls';
+import { fetchSleepLogs } from '../Functions/Functions';
+import { sleepLogsProviderContext } from '../providers/SleepLogsProvider';
 
 type SleepLogManipulateType = {
   id: number,
-  fetchSleepLogs: () => void
 }
 
-export const SleepLogManipulate = ({ id, fetchSleepLogs } : SleepLogManipulateType): JSX.Element => {
+export const SleepLogManipulate = ({ id } : SleepLogManipulateType): JSX.Element => {
+  const { setSleepLogs, setSleepLogsData } = useContext(sleepLogsProviderContext);
   // JSXの表示・非表示を変更する
   const [show, setShow] = useState<boolean>(false);
   const manipulateElement = useRef(null);
   // useCallbackがshowに依存しないよう、Refで新たなStateを持たせる。
   const isShow = useRef(false);
 
+  // 確認したら該当のログを削除して、再度ログを取得する。
   const deleteRequest = () => {
     if (window.confirm("本当に削除しますか?")){
       axios.delete(`${ sleepLogsAPI }/${ id }`, { withCredentials: true })
       .then(res => {
-        fetchSleepLogs();
+        fetchSleepLogs(setSleepLogs, setSleepLogsData);
         console.log(res);
       })
       .catch(e => console.log(e));
